requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0
Pillow>=9.0.0

# Geospatial
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Data Acquisition
requests>=2.31.0
beautifulsoup4>=4.12.0
netCDF4>=1.6.0
xarray>=2023.1.0
cdsapi>=0.5.0

# Deep Learning (Optional)
torch>=2.0.0
torchvision>=0.15.0

# Web Framework (Optional)
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.28.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
numba>=0.57.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Agent-based Modeling (Optional)
mesa>=2.1.0

# Date/Time
pytz>=2023.3
